extract traits Stage and StageExt to reth_stages_api
