+++ bin/reth/src/commands/stage/dump/execution.rs
@@ -10,8 +10,8 @@ use reth_node_ethereum::EthEvmConfig;
 use reth_primitives::stage::StageCheckpoint;
 use reth_provider::{ChainSpecProvider, ProviderFactory};
 use reth_revm::EvmProcessorFactory;
-use reth_stages::{stages::ExecutionStage, Stage};
-use reth_stages_api::UnwindInput;
+use reth_stages::stages::ExecutionStage;
+use reth_stages_api::{Stage, UnwindInput};
 use tracing::info;
 
 pub(crate) async fn dump_execution_stage<DB: Database>(
