+++ bin/reth/src/commands/stage/dump/hashing_storage.rs
@@ -5,8 +5,8 @@ use reth_db::{database::Database, table::TableImporter, tables, DatabaseEnv};
 use reth_node_core::dirs::{ChainPath, DataDirPath};
 use reth_primitives::stage::StageCheckpoint;
 use reth_provider::ProviderFactory;
-use reth_stages::{stages::StorageHashingStage, Stage};
-use reth_stages_api::UnwindInput;
+use reth_stages::stages::StorageHashingStage;
+use reth_stages_api::{Stage, UnwindInput};
 use tracing::info;
 
 pub(crate) async fn dump_hashing_storage_stage<DB: Database>(
