+++ bin/reth/src/commands/stage/dump/merkle.rs
@@ -7,14 +7,11 @@ use reth_node_core::dirs::{ChainPath, DataDirPath};
 use reth_node_ethereum::EthEvmConfig;
 use reth_primitives::{stage::StageCheckpoint, BlockNumber, PruneModes};
 use reth_provider::ProviderFactory;
-use reth_stages::{
-    stages::{
-        AccountHashingStage, ExecutionStage, ExecutionStageThresholds, MerkleStage,
-        StorageHashingStage, MERKLE_STAGE_DEFAULT_CLEAN_THRESHOLD,
-    },
-    Stage,
+use reth_stages::stages::{
+    AccountHashingStage, ExecutionStage, ExecutionStageThresholds, MerkleStage,
+    StorageHashingStage, MERKLE_STAGE_DEFAULT_CLEAN_THRESHOLD,
 };
-use reth_stages_api::UnwindInput;
+use reth_stages_api::{Stage, UnwindInput};
 use tracing::info;
 
 pub(crate) async fn dump_merkle_stage<DB: Database>(
