+++ crates/stages/benches/criterion.rs
@@ -12,9 +12,8 @@ use reth_primitives::{stage::StageCheckpoint, BlockNumber};
 use reth_stages::{
     stages::{MerkleStage, SenderRecoveryStage, TransactionLookupStage},
     test_utils::TestStageDB,
-    Stage, StageExt,
 };
-use reth_stages_api::{ExecInput, UnwindInput};
+use reth_stages_api::{ExecInput, Stage, StageExt, UnwindInput};
 use std::{ops::RangeInclusive, sync::Arc};
 
 mod setup;
