+++ crates/stages/src/stages/index_storage_history.rs
@@ -1,5 +1,4 @@
 use super::{collect_history_indices, load_history_indices};
-use crate::Stage;
 use reth_config::config::EtlConfig;
 use reth_db::{
     database::Database,
@@ -15,7 +14,7 @@ use reth_primitives::{
 use reth_provider::{
     DatabaseProviderRW, HistoryWriter, PruneCheckpointReader, PruneCheckpointWriter,
 };
-use reth_stages_api::{ExecInput, ExecOutput, StageError, UnwindInput, UnwindOutput};
+use reth_stages_api::{ExecInput, ExecOutput, Stage, StageError, UnwindInput, UnwindOutput};
 use std::fmt::Debug;
 use tracing::info;
