+++ crates/stages/src/test_utils/runner.rs
@@ -1,9 +1,10 @@
 use super::TestStageDB;
-use crate::{Stage, StageExt};
 use reth_db::{test_utils::TempDatabase, DatabaseEnv};
 use reth_interfaces::db::DatabaseError;
 use reth_provider::ProviderError;
-use reth_stages_api::{ExecInput, ExecOutput, StageError, UnwindInput, UnwindOutput};
+use reth_stages_api::{
+    ExecInput, ExecOutput, Stage, StageError, StageExt, UnwindInput, UnwindOutput,
+};
 use std::sync::Arc;
 use tokio::sync::oneshot;
