+++ crates/stages/src/test_utils/stage.rs
@@ -1,8 +1,7 @@
-use crate::Stage;
 use reth_db::database::Database;
 use reth_primitives::stage::StageId;
 use reth_provider::DatabaseProviderRW;
-use reth_stages_api::{ExecInput, ExecOutput, StageError, UnwindInput, UnwindOutput};
+use reth_stages_api::{ExecInput, ExecOutput, Stage, StageError, UnwindInput, UnwindOutput};
 use std::collections::VecDeque;
 
 #[derive(Debug)]
